+++ code/packages/sdk/src/actions/images-tool/images-tool.ts
@@ -117,11 +117,10 @@ export class WeaveImagesToolAction extends WeaveAction {
 
       if (dragProperties && dragId === WEAVE_IMAGES_TOOL_ACTION_NAME) {
         this.instance.getStage().setPointersPositions(e);
-        const position: Konva.Vector2d | null | undefined = this.instance
-          .getStage()
-          .getRelativePointerPosition();
 
-        if (!position) {
+        const { mousePoint, container } = this.instance.getMousePointer();
+
+        if (!mousePoint) {
           return;
         }
 
@@ -130,7 +129,8 @@ export class WeaveImagesToolAction extends WeaveAction {
           {
             type: WEAVE_IMAGES_TOOL_UPLOAD_TYPE.IMAGE_URL,
             images: dragProperties.imagesURL,
-            position,
+            container: container as Konva.Layer | Konva.Group | undefined,
+            position: mousePoint,
             ...(dragProperties.forceMainContainer && {
               forceMainContainer: dragProperties.forceMainContainer,
             }),
@@ -377,7 +377,7 @@ export class WeaveImagesToolAction extends WeaveAction {
     const { mousePoint, container } = this.instance.getMousePointer(position);
 
     this.clickPoint = mousePoint;
-    this.container = container as Konva.Layer | Konva.Group;
+    this.container = this.container ?? (container as Konva.Layer | Konva.Group);
 
     const originPoint = {
       x: this.clickPoint?.x ?? 0,
@@ -485,10 +485,8 @@ export class WeaveImagesToolAction extends WeaveAction {
               },
               uploadImageFunction: uploadImageFunctionInternal,
               ...(imageId && { imageId }),
-              position: {
-                x: position.x,
-                y: position.y,
-              },
+              position,
+              container: this.container,
               forceMainContainer: this.forceMainContainer,
               nodeId,
             },
@@ -543,10 +541,8 @@ export class WeaveImagesToolAction extends WeaveAction {
               },
               ...(imageId && { imageId }),
               ...(options && { options }),
-              position: {
-                x: position.x,
-                y: position.y,
-              },
+              position,
+              container: this.container,
               forceMainContainer: this.forceMainContainer,
               nodeId,
             },
@@ -604,6 +600,10 @@ export class WeaveImagesToolAction extends WeaveAction {
 
     if (params?.position) {
       this.setState(WEAVE_IMAGES_TOOL_STATE.SELECTED_POSITION);
+    }
+
+    if (params?.container) {
+      this.container = params.container;
     }
 
     this.nodesIds = [];